Reorganize test directory structure
- Create tests/test_helpers/ for shared test utilities
- Create tests/unit_tests/ for unit tests (moved broadcaster/)
- Create tests/integration_tests/ for integration tests (moved blender/ and blender_manual/)
- Update all import paths to use new absolute paths
- Fix pytest compatibility issues in test classes
- Update CONTRIBUTING.md with testing guidelines for different test types
- Update pyproject.toml test configuration
