+++ tests/integration_tests/blender/test_bpy_library.py
