+++ tests/integration_tests/blender/test_mixer_uv_bpy.py
@@ -1,38 +1,35 @@
 import pytest
-# Create a pytest-compatible wrapper for MixerTestCase
-from tests.mixer_testcase import MixerTestCase as BaseMixerTestCase
-from tests.mixer_testcase import BlenderDesc
+# Create a pytest-compatible wrapper for MixerTestCase using uv bpy
+from tests.test_helpers.mixer_testcase import MixerTestCase as BaseMixerTestCase
+from tests.test_helpers.mixer_testcase import BlenderDesc
 from tests import files_folder
 
+class MixerTestCaseUvBpy(BaseMixerTestCase):
+    """Pytest-compatible wrapper for MixerTestCase using uv bpy"""
 
-class MixerTestCaseWrapper(BaseMixerTestCase):
-    """Pytest-compatible wrapper for MixerTestCase"""
-
-    def __init__(self):
+    def __init__(self, methodName='runTest'):
         # Initialize the base class directly
-        super().__init__()
+        super().__init__(methodName)
         # Add pytest-compatible assertion methods
         self.failureException = AssertionError
 
-    def setup_method(self):
-        # Call the base setup_method with Blender configurations
+    def setup_method(self, method):
+        # Call the base setup_method with Blender configurations using uv bpy
         sender_blendfile = files_folder() / "empty.blend"
         receiver_blendfile = files_folder() / "empty.blend"
         sender = BlenderDesc(load_file=sender_blendfile, wait_for_debugger=False)
         receiver = BlenderDesc(load_file=receiver_blendfile, wait_for_debugger=False)
         blenderdescs = [sender, receiver]
-        super().setup_method(blenderdescs=blenderdescs, join=True)
+        super().setup_method(blenderdescs=blenderdescs, join=True, use_uv_bpy=True)
 
-    def teardown_method(self):
+    def teardown_method(self, method):
         # Clean up blender instances
         super().teardown_method()
 
-
-class TestCase(MixerTestCaseWrapper):
+class TestCaseUvBpy(MixerTestCaseUvBpy):
     pass
 
-
-class TestAnimationData(TestCase):
+class TestAnimationDataUvBpy(TestCaseUvBpy):
     def test_animation_data_clear(self):
         action = """
 import bpy
@@ -58,8 +55,7 @@ obj.hide_viewport = False
 
         self.end_test()
 
-
-class TestKeyFrame(TestCase):
+class TestKeyFrameUvBpy(TestCaseUvBpy):
     def test_create_keyframe_datablock(self):
         action = """
 import bpy
@@ -110,8 +106,7 @@ obj.modifiers[0].keyframe_delete("count")
 """
         self.send_string(action)
 
-
-class TestDriver(TestCase):
+class TestDriverUvBpy(TestCaseUvBpy):
     def test_driver_add(self):
         action = """
 import bpy
@@ -185,7 +180,6 @@ bpy.context.view_layer.update()
         self.send_string(action)
 
         self.end_test()
-
 
 if __name__ == "__main__":
     pytest.main([__file__])
