+++ tests/integration_tests/blender/test_uv_bpy_integration.py
