+++ tests/integration_tests/blender_manual/files/file1.blend
