+++ tests/test_helpers/blender_snippets.py
